+++ packages/@storefront/breadcrumbs/src/breadcrumbs/index.ts
@@ -3,10 +3,11 @@ import { configurable, provide, tag, Events, Selectors, Store, StoreSections, Ta
 @configurable
 @provide<Breadcrumbs.Props, Breadcrumbs.State>(
   'breadcrumbs',
-  ({ showLabels, labels }, { fields, originalQuery, correctedQuery }) => ({
+  ({ showLabels, labels }, { fields, selectedNavigations, originalQuery, correctedQuery }) => ({
     showLabels,
     labels,
     fields,
+    selectedNavigations,
     originalQuery,
     correctedQuery,
   })
@@ -33,16 +34,25 @@ class Breadcrumbs {
     switch (this.props.storeSection) {
       case StoreSections.PAST_PURCHASES:
         this.subscribe(Events.PAST_PURCHASE_SELECTED_REFINEMENTS_UPDATED, this.updateFields);
+        this.subscribe(Events.PAST_PURCHASE_NAVIGATIONS_UPDATED, this.updateFields);
+        this.subscribe(Events.PAST_PURCHASE_NAVIGATIONS_UPDATED, this.updateSelectedNavigations);
         navigationsSelector = () => this.select(Selectors.pastPurchaseNavigations);
         break;
       case StoreSections.SEARCH:
         this.subscribe(Events.ORIGINAL_QUERY_UPDATED, this.updateOriginalQuery);
         this.subscribe(Events.CORRECTED_QUERY_UPDATED, this.updateCorrectedQuery);
         this.subscribe(Events.NAVIGATIONS_UPDATED, this.updateFields);
+        this.subscribe(Events.NAVIGATIONS_UPDATED, this.updateSelectedNavigations);
         navigationsSelector = () => this.select(Selectors.navigations);
         break;
     }
-    this.state = { navigationsSelector, fields: this.getFields(navigationsSelector()), originalQuery: this.select(Selectors.query) };
+
+    this.state = {
+      navigationsSelector,
+      fields: this.getFields(navigationsSelector()),
+      originalQuery: this.select(Selectors.query),
+      selectedNavigations: this.getSelectedNavigations(navigationsSelector())
+    };
   }
 
   onBeforeMount() {
@@ -58,12 +68,26 @@ class Breadcrumbs {
 
   updateCorrectedQuery = (correctedQuery: string) => this.set({ correctedQuery });
 
+  /**
+   * @deprecated
+   */
   updateFields = () => this.set({ fields: this.getFields(this.state.navigationsSelector()) });
 
+  updateSelectedNavigations = () =>
+    this.set({ selectedNavigations: this.getSelectedNavigations(this.state.navigationsSelector()) })
+
+  /**
+   * @deprecated
+   */
   getFields(navigations: Store.Navigation[]) {
     return navigations
       .filter((navigation) => navigation.selected.length !== 0)
       .map((navigation) => navigation.field);
+  }
+
+  getSelectedNavigations(navigations: Store.Navigation[]) {
+    return navigations
+      .filter((navigation) => navigation.selected.length !== 0);
   }
 }
 
@@ -81,6 +105,7 @@ namespace Breadcrumbs {
   export interface State {
     fields: string[];
     originalQuery: string;
+    selectedNavigations: Store.Navigation[];
     correctedQuery?: string;
     navigationsSelector?: () => Store.Navigation[];
   }